Fixed dependency issues

+++ hq_decision_making/include/hq_decision_making/hq_EventSystem.h
@@ -289,7 +289,7 @@ public:
 		void spin(double rate = EQ_SPINNER_DIF_RATE){
 			while(events.check_external_ok() and not events.isTerminated()){
 				spinOne();
-				boost::this_thread::sleep(boost::posix_time::milliseconds(1000/rate));
+				boost::this_thread::sleep(boost::posix_time::milliseconds(int(1000/rate)));
 			}
 			if(not events.check_external_ok() and not events.isTerminated()){
 				events.close();
@@ -306,7 +306,7 @@ public:
 	void spinOne(){ Spinner s(*this); s.spinOne(); }
 	void spin(double rate=EQ_SPINNER_DIF_RATE){ Spinner s(*this); s.spin(rate); }
 	void async_spin(double rate=EQ_SPINNER_DIF_RATE, double start_delay=0.0){
-		boost::this_thread::sleep(boost::posix_time::seconds(start_delay));
+		boost::this_thread::sleep(boost::posix_time::seconds(int(start_delay)));
 		_spinner = boost::shared_ptr<Spinner>(new Spinner (*this));
 		_spinner->start(rate);
 	}
